Support for Java 9+ in `jvmversion` attribute
No need to parse the JVM version as double anymore as all the 1.3 features for which a version number comparison was needed are removed now.

If a version string is passed which starts with '1.' the version is parsed as double and checked that it isn't lower than 1.4, otherwise a warning is issued and the default '1.4+' is used instead.